Optimize herringbone computation
We don't have to store actual angles for ordering we can use their
cosine as well as it's a monotonic function.
An cos may be computed using a simple dot product of normalized vectors.
It's enough to precompute the normalized vectors between all pairs of portals
to make it fast