Fix a potential race condition when fetching a map tile
We clear the missingTiles set and then fill it with a given coord
only after we schedule it to be fetched. In an unfavourable case it may
cause the request to be dropped before we add it to the missingTiles set.
So, instead of clearing it fill it with the necessary tiles, and later
remove some of them if they proof already fetched (in cache).