Fix sizing of the portallist widget
For some reason FLTK was expanding the table widget outsize of the visible area
which was messing up with sizing of the columns ans was hiding the scrollbar.
Add an empty group which will be resizing instead of the portal list.

Also redraw the portallist whenever the tab changes, as the labels of the portals
may need to be redrawn.